# backend/routes/carRoutes.js
const mongoose = require("mongoose");
const Car = mongoose.model("cars");
const requireLogin = require("../middlewares/requireLogin");


function isEmpty(obj) {
  for(var key in obj) {
      if(obj.hasOwnProperty(key))
          return false;
  }
  return true;
}

module.exports = app => {
  app.get("/api/cars", async (req, res) => {
    // ใช้สำหรับหน้าดูรถทั้งหมด
    console.log("query", req.query);
    if (req.query.fromDate !== 'null' && req.query.toDate !== 'null' && !isEmpty(req.query)) {
      console.log('date query')
        const car = await Car.find({
             
                "availFrom": {
                  $lte: req.query.fromDate
                },
                "availTo": {
                    $gte: req.query.toDate
                  },
              
        })
        res.send(car)
    }
      else {
        console.log('normal query')
        const car = await Car.find({});
        res.send(car);
      }
      
    }
  );
  app.post("/api/cars", requireLogin, async (req, res) => {
    // ใช้สำหรับสร้างรถคันใหม่ โดยเจ้าของรถต้องยืนยันตัวตนแล้ว
    const {
      brand,
      type,
      regYear,
      LNumber,
      gear,
      seat,
      equipment,
      photo,
      availFrom,
      availTo,
      description,
      pricePerDay,
      deposit
    } = req.body;
    const car = await new Car({
      brand,
      type,
      regYear,
      LNumber,
      gear,
      seat,
      equipment,
      photo,
      availFrom,
      availTo,
      description,
      pricePerDay,
      deposit,
      _owner: req.user
    }).save();
    console.log(car);
    res.send(car);
  });
  app.get("/api/cars/:id", async (req, res) => {
    // ใช้สำหรับดูรถแต่ละคันได้ โดยใช้ key เป็น id
    const car = await Car.findById(req.params.id).populate('_owner');
    console.log(car)
    res.send(car);
  });
  app.get("/api/ownercars", async (req, res) => {
    // ใช้สำหรับจัดการรถของเจ้าของรถ
    const car = await Car.find({_owner: req.user["_id"]});
    res.send(car);
  })
  app.delete("/api/cars/:id", requireLogin, async (req, res) => {
    // ใช้สำหรับลบรถ
    Car.findByIdAndRemove(req.params.id, (err, blog) => {
      if (err) return res.status(500).send(err);
      const response = {
        message: "Successfully deleted",
        id: Car.id
      };
      return res.status(200).send(response);
    });
  });
};