# backend/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Asset Service",
    "description": "Asset Service"
  },
  "host": "127.0.0.1:5000",
  "basePath": "",
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
    
  },
  "paths": {
    "/api/cars/": {
        "get": {
          "tags": ["car"],
          "summary": "Get all car in database",
          "description": "",
          "operationId": "API for car in the system",
          "consumes": ["application/json", "application/xml"],
          "produces": ["application/xml", "application/json"],
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        }
      },
    "/api/cars/{id}": {
      "get": {
        "tags": ["car"],
        "summary": "Get car in database",
        "description": "",
        "operationId": "API for car in the system",
        "consumes": ["application/json", "application/xml"],
        "produces": ["application/xml", "application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Car id"
          },
          {
            "name": "availFrom",
            "in": "query"
          },
          {
            "name": "availTo",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/auth/login": {
        "post": {
          "tags": ["user"],
          "summary": "Log in",
          "description": "",
          "operationId": "API for user in the system",
          "consumes": ["application/json", "application/xml"],
          "produces": ["application/xml", "application/json"],
          "parameters": [
            {
              "name": "username",
              "in": "query"
            },
            {
              "name": "password",
              "in": "query"
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        }
      },
      "/auth/local": {
        "post": {
          "tags": ["user"],
          "summary": "Sign up",
          "description": "",
          "operationId": "API for user in the system",
          "consumes": ["application/json", "application/xml"],
          "produces": ["application/xml", "application/json"],
          "parameters": [
            {
              "name": "username",
              "in": "query"
            },
            {
              "name": "password",
              "in": "query"
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        }
      },
      "/api/request": {
        "get": {
          "tags": ["request"],
          "security": [
            {
              "Bearer": []
            }
          ],
          "summary": "Get all request in database",
          "description": "",
          "operationId": "API for request in the system",
          "consumes": ["application/json"],
          "produces": ["application/json"],
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        }
      },
      "/api/request/{requestId}": {
        "put": {
          "tags": ["request"],
          "summary": "Using to complete the request",
          "description": "",
          "operationId": "API for request in the system",
          "consumes": ["application/json", "application/xml"],
          "produces": ["application/xml", "application/json"],
          "parameters": [
            {
              "name": "requestId",
              "in": "path"
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        },
        "delete": {
          "tags": ["request"],
          "summary": "Delete the request",
          "description": "",
          "operationId": "API for request in the system",
          "consumes": ["application/json", "application/xml"],
          "produces": ["application/xml", "application/json"],
          "parameters": [
            {
              "name": "requestId",
              "in": "path"
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        }
      }
  },
  "definitions": {
    "availFrom": {
      "type": "Date"
    },
    "availTo": {
        "type": "Date"
      }
  }
}